concurent runners seem to be working now!
The issue came from my understanding on how to read `stdout`. I need to read the output and *then* wait for the process. The blocking part is done by reading the output which copies to a buffer until it reaches EOF.

Other bugs remains (but they are related to the encoding this time)