I hate myself
Somewhere the there is a clash of variable id. This wasn't a problem before because they also use their sort to generate unique names, and I suspect I abused that fact. However with the new `cryptovampire_smt` this is no longer the case.
This commit adds a new way to generate the id of a varaible for smt using a bit of bitwise operations